fix(home): address additional PR review feedback
- Remove invalid PageSection variant="light" prop
- Conditionally show "Learn more" column only when documentationUrl is set
- Merge HomeSectionCard, DashboardIngestionMetrics, and
  PortfolioMetricsSection into a single file
- Remove invalid Card isFlat prop
- Default latestSbom/latestAdvisory to null when undefined
